Better namespace/folder grouping
