+++ src/Tingle.EventBus.Transports.InMemory/Client/BroadcastChannelWriter.cs
@@ -5,9 +5,9 @@ using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
 
-namespace Tingle.EventBus.Transports.InMemory
+namespace Tingle.EventBus.Transports.InMemory.Client
 {
-    class BroadcastChannelWriter<T> : ChannelWriter<T>
+    internal class BroadcastChannelWriter<T> : ChannelWriter<T>
     {
         public BroadcastChannelWriter(ICollection<ChannelWriter<T>>? children = null)
         {
