+++ src/Tingle.EventBus.Transports.InMemory/Client/InMemoryMessage.cs
@@ -1,30 +1,33 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Tingle.EventBus.Transports.InMemory
+namespace Tingle.EventBus.Transports.InMemory.Client
 {
     /// <summary>
-    /// The <see cref="InMemoryMessage"/> is used to receive data from InMemory queues and subscriptions.
+    /// The <see cref="InMemoryMessage"/> is used to receive
+    /// and send data from and to InMemory entities.
     /// </summary>
-    public class InMemoryReceivedMessage
+    public class InMemoryMessage
     {
+        /// <summary>
+        /// Creates a new message from the specified string, using UTF-8 encoding.
+        /// </summary>
+        /// <param name="body">The payload of the message as a string.</param>
+        public InMemoryMessage(string body) : this(BinaryData.FromString(body)) { }
+
+        /// <summary>
+        /// Creates a new message from the specified payload.
+        /// </summary>
+        /// <param name="body">The payload of the message in bytes.</param>
+        public InMemoryMessage(ReadOnlyMemory<byte> body) : this(BinaryData.FromBytes(body)) { }
+
         /// <summary>
         /// Creates a new message from specified <see cref="BinaryData"/> instance.
         /// </summary>
         /// <param name="body">The payload of the message.</param>
-        internal InMemoryReceivedMessage(BinaryData body)
+        public InMemoryMessage(BinaryData body)
         {
             Body = body ?? throw new ArgumentNullException(nameof(body));
-        }
-
-        internal InMemoryReceivedMessage(InMemoryMessage message) : this(message?.Body ?? throw new ArgumentNullException(nameof(message)))
-        {
-            ContentType = message.ContentType;
-            CorrelationId = message.CorrelationId;
-            MessageId = message.MessageId;
-            SequenceNumber = message.SequenceNumber;
-            Properties = message.Properties;
-            Scheduled = message.Scheduled;
         }
 
         /// <summary>
@@ -35,7 +38,7 @@ namespace Tingle.EventBus.Transports.InMemory
         /// Optionally describes the payload of the message, with a descriptor
         /// following the format of RFC2045, Section 5, for example "application/json".
         /// </remarks>
-        public string? ContentType { get; }
+        public string? ContentType { get; set; }
 
         /// <summary>
         /// Gets or sets the a correlation identifier.
@@ -46,7 +49,7 @@ namespace Tingle.EventBus.Transports.InMemory
         /// purposes of correlation, for example reflecting the <see cref="MessageId"/>
         /// of a message that is being replied to.
         /// </remarks>
-        public string? CorrelationId { get; }
+        public string? CorrelationId { get; set; }
 
         /// <summary>
         /// Gets or sets the MessageId to identify the message.
@@ -57,12 +60,12 @@ namespace Tingle.EventBus.Transports.InMemory
         /// string and can reflect a GUID or an identifier derived from the
         /// application context.
         /// </remarks>
-        public string? MessageId { get; }
+        public string? MessageId { get; set; }
 
         /// <summary>
         /// Gets or sets the body of the message.
         /// </summary>
-        public BinaryData Body { get; }
+        public BinaryData Body { get; set; }
 
         /// <summary>
         /// Gets the unique number assigned to a message by the transport.
@@ -73,7 +76,7 @@ namespace Tingle.EventBus.Transports.InMemory
         /// Sequence numbers monotonically increase. They roll over to 0 when the 48-64 bit
         /// range is exhausted. This property is read-only.
         /// </remarks>
-        public long SequenceNumber { get; }
+        public long SequenceNumber { get; internal set; }
 
         /// <summary>
         /// Gets the application properties bag, which can be used for custom message metadata.
@@ -89,6 +92,6 @@ namespace Tingle.EventBus.Transports.InMemory
         /// Gets the date and time in UTC at which the message will be enqueued.
         /// This property returns the time in UTC.
         /// </summary>
-        public DateTimeOffset Scheduled { get; }
+        public DateTimeOffset Scheduled { get; set; }
     }
 }
