+++ src/Tingle.EventBus.Transports.InMemory/Client/InMemoryProcessorOptions.cs
