+++ src/Tingle.EventBus.Transports.InMemory/Client/InMemorySender.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
 
-namespace Tingle.EventBus.Transports.InMemory
+namespace Tingle.EventBus.Transports.InMemory.Client
 {
     internal sealed class InMemorySender : IDisposable
     {
