+++ src/Tingle.EventBus.Transports.InMemory/Client/SequenceNumberGenerator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading;
 
-namespace Tingle.EventBus.Transports.InMemory
+namespace Tingle.EventBus.Transports.InMemory.Client
 {
     /// <summary>
     /// A sequence number generator for InMemory transport
