+++ src/Tingle.EventBus.Transports.InMemory/EventBusBuilderExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Options;
 using System;
 using Tingle.EventBus.Transports.InMemory;
+using Tingle.EventBus.Transports.InMemory.Client;
 
 namespace Microsoft.Extensions.DependencyInjection
 {
